requirements.txt
kafka-python
pandas
faker
snowflake-connector-python

